+++ platform/android/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".HomeScreenActivity"
+            android:name=".HomeScreenFragment"
             android:exported="false"
             android:label="@string/title_activity_home_screen"
             android:theme="@style/AppTheme" />
