+++ platform/android/app/src/main/kotlin/coop/polypoly/polypod/MainActivity.kt
@@ -18,7 +18,6 @@ class MainActivity : AppCompatActivity() {
     }
 
     private var onboardingShown = false
-    private val featureStorage = FeatureStorage()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -42,7 +41,6 @@ class MainActivity : AppCompatActivity() {
 
         Authentication.authenticate(this) { success ->
             if (success) {
-                featureStorage.importFeatures(this)
                 setContentView(R.layout.activity_main)
                 setSupportActionBar(findViewById(R.id.toolbar))
             } else {
