+++ platform/android/app/src/main/res/layout/activity_main.xml
@@ -13,7 +13,7 @@
 
         <androidx.fragment.app.FragmentContainerView
             android:id="@+id/fragmentContainerView"
-            android:name="coop.polypoly.polypod.HomeScreenActivity"
+            android:name="coop.polypoly.polypod.HomeScreenFragment"
             android:layout_width="0dp"
             android:layout_height="0dp"
             app:layout_constraintBottom_toBottomOf="parent"
